# Curl improvements