--- a/curl/common.go
+++ b/curl/common.go
@@ -4,21 +4,12 @@ import (
 	. "github.com/iotaledger/iota.go/consts"
 )
 
-// CurlRounds is the default number of rounds used in transform.
-type CurlRounds int
-
 const (
 	// StateSize is the size of the Curl hash function.
 	StateSize = HashTrinarySize * 3
 
-	// CurlP27 is used for hashing with 27 rounds
-	CurlP27 CurlRounds = 27
-
-	// CurlP81 is used for hashing with 81 rounds
-	CurlP81 CurlRounds = 81
-
-	// NumberOfRounds is the default number of rounds in transform.
-	NumberOfRounds = CurlP81
+	// NumRounds is the number of rounds in a Curl transform.
+	NumRounds = 81
 )
 
 // SpongeDirection indicates the direction trits are flowing through the sponge.