--- a/curl/transform_decl.go
+++ b/curl/transform_decl.go
@@ -1,7 +1,0 @@
-// +build !gccgo,!appengine
-// +build amd64
-
-package curl
-
-//go:noescape
-func transform(dst, src *[StateSize]int8, rounds uint)