--- a/curl/transform_generic.go
+++ b/curl/transform_generic.go
@@ -1,7 +1,0 @@
-// +build !amd64 appengine gccgo
-
-package curl
-
-func transform(dst, src *[StateSize]int8, rounds uint) {
-	transformGeneric(dst, src, rounds)
-}